BUG: Allow projections of analytic columns
This seems to fix https://github.com/cloudera/ibis/issues/900
Doesn't seem to be a very satisfactory fix because it removes a bunch
of logic. All the tests still pass though.    I'm unsure of the
original intentions of that logic because the if exist_layers: was
never able to evaluate True as it's only set to True in the while
loop. The while loop has non changing conditions so it would just
error out if the loop was entered.

Author: Max Mizikar <maxmzkr@deepfield.net>

Closes #901 from maxmzkr/fix-analytic-projection and squashes the following commits:

256b3e7 [Max Mizikar] Test that anyalytics with a filter compile
7c0f03f [Max Mizikar] Test analytic selection ancestor
b3a7591 [Max Mizikar] Allow projections of analytic columns