ENH: Add a data_preload step in pandas backend
This PR adds a `data_preload` step to the execution pipeline for the
Pandas  backend.    The motivation for this is to allow clients to
perform operations on `(Node,  ConcreteData)` pairs (including a
`scope` argument) *before* the execution  for a particular operation
starts.    Custom data source execution is the main use case here.
As a motivating example, consider a custom data source object that can
be  turned into a DataFrame.    We want to be able to operate on our
custom data source, but we don't want to  redefine every operation for
this data source, or it doesn't make sense to  define operations on
this custom object directly.    `data_preload` gives the ability to
call a function on data pieces that are in scope before any execution
happens. In the example above, `data_preload` would turn the
`ConcreteData` object into a `pandas.DataFrame`.    The default
behavior is no-op.    This also sets the stage for multi-client
execution. More on that to come.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1142 from cpcloud/data-preload and squashes the following commits:

11124fd [Phillip Cloud] ENH: Add a data_preload step in pandas backend