BUG: Fix subqueries with parameters
closes #1300   closes #1331     This was a somewhat invasive change,
but nets a lot fewer places where the `params` argument has to be used
outside of the `compile` and `execute` functions and methods on
clients. Now, `params` are part of the `context` and `context` is
never `None` so it's easier to reason about where parameters are
coming from.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1303 from cpcloud/fix-scalar-params and squashes the following commits:

eb285ded [Phillip Cloud] BUG: Fix scalar parameters