BUG/CI: Insert chunksize for sqlite should be 1
Pandas 0.23 changed the default behavior of the `chunksize` argument
which breaks SQLite when loading a large dataset because SQLite has an
upper bound on the number of rows in a single insert statement.

This PR works around that our `ci/datamgr.py` script so that we can get
master back to green.