Use weakref.WeakSet to hold ImpalaTemporaryTable instances
We were incorrectly using weakref.WeakValueDictionary to hold temp table
instances that mapped to themselves.

This defeats the purpose of using a weakref container because the
WeakValueDictionary will hold a strong reference to the key and weak
reference to the value. In this scenario we will forever hold at least
one strong reference per key and leak memory.

By using weakref.WeakSet we now only weak references on the table
instance.