Test for Alchemy-based full outer join
* Test using pytest framework that would catch bug in issue #1772
(outer_join actually did a left join).
* Fixed outer join portion of test_joins test method.
* Removed unittest version of full outer test.
* Refactored MockConnection and MockAlchemyConnection.

In order for the test to catch this problem, it required changes to the
MockAlchemyConnection class. Previously that class's dialect and
_build_ast methods actually used implementations specific to Impala
(objects from `ibis.impala.compiler`). This meant that the `outer_join`
method compiled correctly in the sqlalchemy test because it was using
the Impala compiler, whose implementation did not have the bug.

To make clearer the similarities between Impala- and Alchemy-flavored
mock connections, I refactored so they both share a BaseMockConnection
class. However, I kept the name of the Impala-flavored one as
MockConnection because it is used in several other tests. I wasn't sure
if those other tests really should use Impala or Alchemy. As a to-do,
those other tests should be checked for which flavor they ought to use,
and the MockConnection should be renamed to MockImpalaConnection.