Postgres UDF testing: fix to exclude pl/python from Windows tests
The Azure Windows pipelines use a version of postgres that does not
include any extensions (e.g. PostGIS or PL/Python), so the testing
infrastructure should not try to create extensions, and it should
exclude tests for them in the Windows pipeline configuration. This
commit makes those changes

* adds command line option to `postgres` on whether to create plpython
extension
* in `load_data.sh` (used in Linux-based test pipelines) specify to
create plpython with the command line option
* adds a new pytest marker `postgres_extensions` to `test_udf.py` and
`test_postgis.py` and to the marker registry in `setup.cfg`
* substitutes `postgres_extensions` marker in exclusions in
`windows.yml` pipeline configuration