Respect dependency of impala on postgres service
The impala service actually relies on there being an instance of
postgres available to use as its metadata store. It has a timer to
check up to 120 seconds and then impala startup hangs indefinitely or
fails.

When we switched to a new image for postgres that includes PL/Python,
this pushed the startup time just long enough to cause impala startup
to sometimes fail (unpredictably).

Instead of increasing the wait time hard-coded in the Impala image,
it seems better/easier to enforce the dependency with the
`docker-compose.yml`.

See https://github.com/ibis-project/docker-impala/blob/master/bin/start-impala.sh line 21 at revision dcb7388

`/wait-for-it.sh postgres:5432 -t 120`