ENH: Validate AsOfJoin tolerance and attempt interval unit conversion
Unlike other operations, not all of the arguments of AsOfJoin are
automatically passed through the validation chain that occurs in
superclass Annotable, since they are not passed into super(). The
result is that whereas other operations with offset-like arguments
(e.g. Lead/Lag) have their offsets' dtypes automatically inferred, the
same does not occur for 'tolerance' in AsOfJoin. We can simply wire
this validation in ourselves.    Additionally, the default unit for
Interval is 's' regardless of the unit of the value of origin. We can
make a better attempt at inferring unit before immediately defaulting
to seconds.
Author: Emily Reff <emily.reff@twosigma.com>

Closes #1952 from emilyreff7/hackday and squashes the following commits:

2d7ff71 [Emily Reff] Merge branch 'master' of https://github.com/ibis-project/ibis into hackday
40436e1 [Emily Reff] break up tests
134caf3 [Emily Reff] fix failing tests
91318be [Emily Reff] fix failing tests
c524401 [Emily Reff] fix failing tests
97fa826 [Emily Reff] validate AsOfJoin tolerance and attempt interval unit conversion