ENH: SQLAlchemy Default precision and scale to decimal types for PostgreSQL and MySQL
Possible fix for #1273     Adds default precision and scale to decimal
types for PostgreSQL and MySQL. Numeric data types without defined
precision and scale are allowed in these dialects (even if it is
recommended that you specify them), and these tables exist in the
wild.    I'm not sure if this is the best fix, but I was hoping to get
a conversation started.
Author: Ian Rose <ian.r.rose@gmail.com>
Author: Ivan Ogasawara <ivan.ogasawara@gmail.com>

Closes #1969 from ian-r-rose/default-precision and squashes the following commits:

70e36d8d [Ian Rose] Merge pull request #2 from Quansight/default-precision
6f584e0f [Ivan Ogasawara] Fixed tests
66ed4703 [Ivan Ogasawara] Add test for decimal default precision and scale
989b266a [Ian Rose] Add tests for default precision and scale for PostgreSQL.
269328bb [Ian Rose] Also register base sqlalchemy dialect.
85bebf72 [Ian Rose] Use specific sqlalchemy dialect types.
47ca505d [Ian Rose] Add default precision and scale to decimal types for PostgreSQL and MySQL.