TST: mark failing BigQuery tests as xfail
It appears that most of these issues are due to changes in how a
DataFrame is created in the `google-cloud-bigquery` library. Now that
Arrow is used as an intermediate format, arrays and structs come back as
numpy arrays or Arrow objects. Previously they came back as Python
objects (lists and dictionaries).