ci(docs): skip backend tests if _only_ docs have been touched
This moves all of the docs + linting checks to a separate GHA yml file
that is always executed for pushes to or PRs against `master`.

Other tests are in `ibis-backends.yml` and `ibis-main.yml` and these are
also run for pushes and PRs except if the only files touched are within
the `docs/` subfolder.