ci(docs): add dummy jobs that return in place of skipped required jobs
These intentionally collide with the jobs in `ibis-main.yml` and
`ibis-backends.yml` and run on the inverse of the path skipping logic
present in those two files.

This should allow CI to finish and all jobs to return, even if they're
skipped.

This is an ugly hack, but is the supported one for now.