test: dynamic fixtures act on backend-specific tests
`conftest.py` files get run for child directories, so we can re-use the
backend xfails in backend specific tests if they are defined in a parent
directory.

So I've moved the contents of `backends/tests/conftest.py` into
`backends/conftest.py` to accomplish this.

I've also added in checks to make sure we don't inadvertantly run a test
with multiple backends simultaneously (this can happen if a
backend-specific con is included alongside a general backend, for instance).

Also require that `backends_never` usage provide a `reason`