docs(sqlite): remove unavailable sql method from sqlite docs
Also adds a `NotImplementedError` to the base class for child classes
that don't implement `_get_schema_using_query`.

This isn't the solution that I wanted for the docs, but it works.

There are options to `filter` out `members` in `mkdocstrings` but the
`filter` option is currently incompatible with `inherited_members`.