feat(ver): dynamic version identifiers
This adds a dynamic version identifier to Ibis installs -- this should
have no discernable impact on users installing from `conda` or
`pypi` (or from released versions generally).

For users who install from source, it generates a version of the form:

```
<most_recent_release>-post.<numcommits_since_release>+<short_sha_of_HEAD>(.dirty)?
```

The `dirty` suffix is added if there are unstaged changes when the
install is performed.
This will help us help users and other devs track which exact version is
installed when troubleshooting / comparing performance.

The `version` identifier in `pyproject.toml` is not affected by this
process and should display the last released version (currently 2.1.1).

The `__version__` string in `ibis/__init__.py` is now also hard-coded to `"2.1.1"`

During a release, `poetry version` will bump the version in the
`pyproject.toml` file.  I've also added a plugin for the semantic
release bot to automatically replace the version string in the
`__init__.py` file.