fix: use bounded precision for decimal aggregations
This tweaks the new behavior around decimal precision and scale.  If a
decimal has unbounded precision/scale, specified as `(None, None)`, then
the precision/scale of an aggregation of that column will also have
unbounded precision.

For a decimal with some bounded precision/scale, e.g. `(18, 2)`, the
precision/scale of an aggreagtion of that column will have a precision
of `(38, 2)` which is the previous default maximum for bounded precision
and matches up with _most_ backends.