test(pyspark): add tests for NaN handling in aggregations
Backends like the Pandas backend treat np.nans as nulls when computing
aggregations (e.g. count). The PySpark backend was not treating np.nans
as nulls, leading to results for aggregations that are inconsistent with
other backends. This change adds tests for checking that np.nans are
filtered out in aggregates.