feat: add to_pyarrow and to_pyarrow_batches
Adds `to_pyarrow` and `to_pyarrow_batches` to the `BaseBackend`.

`to_pyarrow` returns pyarrow objects consistent with the dimension of
the output:
  - a table -> pa.Table
  - a column -> pa.Array
  - a scalar -> pa.Scalar

`to_pyarrow_batches` returns a RecordBatchReader that returns batches of
pyarrow tables.  It does not have the same dimension handling because
that is not available in RecordBatchReaders.

`to_pyarrow_batches` is implemented for `AlchemyBackend`, `datafusion`,
and `duckdb`.

The `pandas` backend has `to_pyarrow` implemented by using
`pandas.DataFrame.to_pyarrow()`.

Backends that do not require pyarrow already will only require it when
using to_pyarrow* methods.

There are warnings on these methods to indicate that they are
experimental and that they may break in the future irrespective of
semantic versioning.

The DuckDB `to_pyarrow_batches` makes use of a proxy object to escape
garbage collection so that the underlying record batches are still
available even after the `cursor` used to generate them _would have_
been garbage collected (but isn't because it is embedded in the proxy
object)