chore(dev-deps): remove `ipdb` from dev-dependencies
This package incorrect requires setuptools at runtime which causes nix
builds to break with an infinite recursion error, because setuptools is
required for bootstrapping. I put up a PR to address this a few days
ago--https://github.com/gotcha/ipdb/pull/254--but it's not clear whether
the project is maintained.