refactor(ir): remove `ops.TopKNode` and `ir.TopK`
These abstractions add a lot of extra code the save the user no lines in
the case of the aggregation version, and a single line in the case of
the filter version, so we remove it.

BREAKING CHANGE: Replace `t[t.x.topk(...)]` with `t.semi_join(t.x.topk(...), "x")`.