refactor(timestamps): remove `timezone` argument to `to_timestamp` API
This argument was only supported in the BigQuery backend, whose
functionality can be acheived by appending a `%Z` to the format string
and appending the desired timezone to the column being parsed.

BREAKING CHANGE: The `timezone` argument to `to_timestamp` is gone. This was only supported in the BigQuery backend. Append `%Z` to the format string and the desired time zone to the input column if necessary.