refactor: remove the `JSONB` type
This type isn't well supported. Currently, its only purpose is to
allow users to consume tables with a postgres-native JSONB type, but the
expression types provide no novel APIs beyond binary operations (which
is an implementation detail of JSONB). This commit removes the JSONB
type and uses the existing JSON type for table consumption.

BREAKING CHANGE: The `JSONB` type is replaced by the `JSON` type.