feat(sqlite): implement `_get_schema_using_query` in SQLite backend
As SQLite does not provide this functionality directly the code here
creates a temporal view for the query and then uses `table_info`
pragma for retrieving the schema.

A convoluted and changing named is used for the view in order to avoid
collisions. The lifetime of SQLAlchemy SQLite backend connections is
not documented so this seems to be the safest thing to do.

The code has been highly inspired by DuckDB backend code.