feat: return Table from create_table(), create_view()
Closes https://github.com/ibis-project/ibis/issues/5693

I didn't add any tests sepcific to this,
because I modified some existing tests so that
they relied on this new behavior, so this new
behavior is getting tested as a side benefit.
The tests are also clearer now because of this.
I did NOT adjust some tests that I thought would
more clear to have two separate steps of
creating the table/view and accessing it later.