fix(api): disambiguate attribute errors from a missing `resolve` method
Previously we caught `AttributeError`s to consider values that aren't
`Deferred` instances as resolved.

However, we caught the `AttributeError` during resolution of the
`resolve` attribute _and_ resolution of the deferred value against an
object.

This commit scopes that to only the `resolve` attribute, other errors
are bubbled up as usual.