chore(deps): bump lower bound of ipython in dev dependencies
We had a contributor run into issues with a very old version of
`prompt_toolkit` (3.0.4, ~3 years old) in the conda-lock files cause
trouble.

This bumps the lower bound of `ipython` (which is the package that pulls
in `prompt_toolkit`) so that we get a less-old version.

I don't think this counts as a breaking change in that it only impacts
dev dependencies.