refactor(ops.Hash): remove `how` from backend-specific hash operation
BREAKING CHANGE: The `hash` method on table columns on longer accepts
the `how` argument.  The hashing functions available are highly
backend-dependent and the intention of the hash operation is to provide
a fast, consistent (on the same backend, only) integer value.
If you have been passing in a value for `how`, you can remove it and you
will get the same results as before, as there were no backends with
multiple hash functions working.