fix(pandas): preserve RHS values in asof join when column names collide
`pandas.merge_asof` loses information when any of the grouping or
predicate columns share the same name.  It outputs a single column (e.g.
`time`) instead of two columns (`time`, `time_right`).

Ibis attempts to recover the now "missing" column, but previously, we
were simply creating a copy of the LHS `time` column and then renaming
it to be `time_right`.

This behavior also occurred in grouping columns if the names match.

Now, we add default join suffixes to predicates and groups and rename the
corresponding columns before the `merge_asof`.