fix(bigquery): allow passing both schema and obj in create_table
The previous implementation would ignore a dataframe or pyarrow table
passed in as `obj` is `schema` was specified.  Nothing would error, but
it bigquery would create an empty table according to the provided
schema.

Now, if both are passed, we create a memtable using the obj and the
schema and pass that to bigquery.

There is still an edge case of what happens if `obj` is an `ir.Table`
and schema is also specified.