fix(postgres): by default use domain socket
postgres:///dbname tries to connect via localhost, which may not be
allowed.  With localhost as the default parameter, there is no way to
force it to use the local unix domain socket.  This would make the
default using the domain socket, and then connecting via localhost
can be used via "postgres://localhost/dbname".