refactor(impala): modernize the impala backend
Remove HDFS integration from the impala for dramatic simplification of the backend.

Also move away from an ancient version of impala that is no longer maintained.

The primary motivation here is to reduce the scope of the Impala backend, and to lower its maintenance by supporting fewer features.

BREAKING CHANGE: Direct HDFS integration is removed and support for ingesting pandas DataFrames directly is as well. The Impala backend still works with HDFS, but data in HDFS must be managed outside of ibis.