fix(deferred): don't pass expression in fstringified error message
This is a weird one, BUT:
- There is (at least) one case where, when trying to resolve a deferred
value, we end up at the `CoercionError` exception with a non-deferred
expression.

When `ibis.options.interactive = True`, we then attempt to render that
expression using `rich` because that's what f-strings do.

Even though in the calling code, which here is in `CoercedTo.match`, we
catch the `CoercionError`, the `CoercionError` never gets thrown because
we've instead errored out trying to render an f-string that will never
be seen.

If, instead of rendering the expression, we instead mention the
expression type, that prevents the short-circuit.