fix(duckdb): allow table creation from expr with geospatial datatypes
I thought, initially, that `WKB_BLOB` should be mapped on to something
like `dt.GeoSpatial`, but that leads to a situation where we try to call
`AsWKB` on an existing WKB column, and DuckDB does not like that.

Mapping it as `binary` works and the created columns are still correctly
detected as `geometry:geospatial`.