refactor(ir): add `JoinTable` operation unique to `JoinChain` instead of using the globally unique `SelfReference`
This enables us to maintain join expression equality:
`a.join(b).equals(a.join(b))`

So far we have been using SelfReference to make join tables unique, but
it was globally unique which broke the equality check above. Therefore
we need to restrict the uniqueness to the scope of the join chain. The
simplest solution for that is to simply enumerate the join tables in
the join chain, hence now all join participants must be
`ops.JoinTable(rel, index)` instances.

`ops.SelfReference` is still required to distinguish between two
identical tables at the API level, but it is now decoupled from the
join internal representation.