feat(api): add `.as_scalar()` method for turning expressions into scalar subqueries (#8350)
Since there is no safe way to determine whether a table or a column
contains a single cell we need to provide a way for the users to inform
ibis that it should treat an expression as a scalar subquery. Previously
it was the `.to_array()` method, but it only worked for table
expressions.

Note that when a scalar reduction is passed to certain API methods ibis
is able to infer that the value should be turned into a scalar subquery
and it implicitly does it. Using this method can be omitted in those
cases.