fix(comparison): wrap isnull equality check in parens
We wrap arguments to a `sge.Equals` in parens if the inputs are
themselves binary operations.  There are other operations we will have
to wrap -- for now I'm just adding `isNull` to the list, but I imagine
this will grow, possibly on a backend-by-backend basis.

I'm going to investigate the mssql and oracle failures but they are
unrelated to this PR (they fail the added test with or without the parentheses)