perf: avoid exponential growth on `name` attribute access
This commit changes the use of `hasattr` + `name` attribute access to
eagerly access the `name` attribute, store it and check for `None`.
Accessing the attribute twice may seem innocuous, but since names may be
generated recursively, this results in 2 ** n growth in the number of
accesses: once for the `hasattr` call, and once for the access if
`hasattr` succeeds. This can happen with large operation trees where
names are automatically generated.