refactor(oracle): remove dependency on private function for temp tables (#8480)
## Description of changes

The previous workaround for temp tables in Oracle was quite brittle and
broke
when I tried to bump our upper pin on `sqlglot`. I've refactored it to
be
(hopefully) less brittle ~and also bumped the `sqlglot` pin up a few
versions.~

Dropping the version bump from this PR, there are some breaking changes
for postgres, risingwave, and trino, looks like.