perf(sql): prevent sqlglot from extensive deepcopying every time we create a sqlglot object (#8592)
Still profiling and adding more `copy=False` options, apparently it
greatly improves the performance.

According to profiling the recursive generation of sqlglot is still a
bottleneck for big queries which cannot be fixed on the ibis side. There
could be one option though to compile the fragments in a greedy fashion
which are going to be cached by `Node.map()` and inject those as
arbitrary strings to other sqlglot expressions.

Now it uses 30% of the memory it used before. Apparently there are still
a lot of sqlglot Literal objects around, looking into that.

fixes #8484

---------

Co-authored-by: Phillip Cloud <417981+cpcloud@users.noreply.github.com>