test(chore): deduplicate and fix marks in test_map
The mark for risingwave was misleading. It was sometimes combined with
the postgres mark, saying the reason was that
only string -> string is supported. But that
isn't true.

I had to get rid of the `raises=PsycoPg2InternalError` in the risingwave mark, because for some tests the
error was actually raised by a lack of the FromYMD Operation. pytest marks don't have support for
"this error is caused by one of these two reasons", so I have to go with the lowest common denominator,
and not list a reason.