feat(mssql): use integrated auth when no user or password supplied
The method for enabling Integrated Authentication with ODBC is to not
provide a `user` or `password`, and to set `Trusted_Connection =
'true'`.

If `Trusted_Connection` is already set in `kwargs`, do nothing, because
either the user set it to 'true' and it will work untouched, or they're
specifying 'false' for a reason and we should leave it alone.