fix(flink): fix compilation of memtable with nested data (#8751)
## Description of changes

This PR aims to fix the compilation of memtables with nested data.

### What was broken

In particular, [Flink does not support the ``STRUCT(1 AS `a`)`` aliasing
syntax to define named
STRUCTs](https://issues.apache.org/jira/browse/FLINK-9161). In order to
do so, we must use a workaround using `CAST`, e.g.,
```sql
SELECT CAST(('a', 1) as ROW<a STRING, b INT>);
```

However, Flink also does not allow you to directly construct ARRAYs of
named STRUCTs using the `ARRAY[]` constructor. This is a bug that I
identified and I have filed it with the Flink community (JIRA ticket
ref: https://issues.apache.org/jira/browse/FLINK-34898).

For the time being, we will need to use another `CAST` workaound that
casts the entire nested array, e.g.,
```sql
SELECT cast(ARRAY[ROW(1)] as ARRAY<ROW<a INT>>);  -- instead of ARRAY[CAST(ROW(1) AS ROW<a INT>)]
```

### How to fix

To summarize,
- if it’s an array of named structs
`CAST(ARRAY[] AS ARRAY<ROW<>, ROW<>>)`
- if it’s named structs
`CAST(ROW() AS ROW<datatype of each field>)`
- if it’s unnamed structs (but I'm not sure how to write this in Ibis)
`ROW()`

I thought of two approaches to this:
1. Rewrite the operator mapping in the Flink backend (i.e., change the
`visit_NonNullLiteral()` method)
2. Rewrite the translation rule in Flink's `Generator`

I found both implementations in different scenarios and decided to go
with option (2).

## Issues closed

#8516