feat(api): add `catalog` and `database` kwargs to `ibis.table`
This allows specifying 1 or 2 levels of table positional hierarchy in a
call to `ibis.table`. I've updated the reprs for `UnboundTable` and
`DatabaseTable` to show that hierarchy when it is specified.

I've also swapped the attribute order for `ops.Namespace` as a
convenience to allow unpacking the argnames in the correct hierarchical
order.