fix(duckdb): disable batching for calls to `arrow`
xref https://github.com/ibis-project/ibis/discussions/8896

This one took me a _while_ to figure out.

the `arrow` method has a `batch_size` kwarg that defaults to 1_000_000,
but the batch sizes aren't actually deterministic, they're just close to
that size. If the result-set is larger than `batch_size`, you can get
slightly different results because of the batching and where the
boundaries are.

I can add a test for this, but it will require a query that returns
north of a million rows and I don't know if we want that in our unit tests.