feat(api): add API for unwrapping JSON values into backend-native values (#8958)
Adds `str`, `int`, `float`, and `bool` properties to `JSONValue` as well
as an `unwrap_as` method for easier programmatic usage and more
fine-grained casting.

Unless someone really hates the static property names, I'd prefer to
keep them as they are. Open to alternative names for `unwrap_as` though.

In theory this can all be done with casting, but if you look at what's
being done in the various backends it's typically a lot more involved
than that. Trino in particular requires queries over JSON to be
`VARCHAR` inputs, when then have to be cast **back** to its `JSON` type
to be able to cast _that_ to the desired output type.

Complicating the cast branching _just_ for the `JSON -> not JSON` case
seemed like the wrong tradeoff.

I went with these names to match the `map` and `array` APIs, and to
match the short type names we have for the specific types (`str`, `int`,
`float`, and `bool`), which exist to match the equivalent Python types.