fix: make devcontainer work correctly (#9019)
## Description of changes

These change are what was required for me to get the devcontainer to run
on my machine. Before these changes, the devcontainer image wasn't
building correctly because it was missing gdal. The root cause of that
was that the devcontainer.json wasn't using the Dockerfile.

- The previous devcontainer.json wasn't actually using the Dockerfile,
now it does
- In devcontainer.json, I switched updateContent command for
postCreateCommand as that looks more appropriate based on the docs
- A couple of minor tidying changes

## Issues closed

Resolves #9011