feat(postgres): add mappings for more esoteric dtypes (#9055)
Resolves #8845

Postgres' `information_schema` has a few custom dtypes that are only
used there. Most of them just map to stringlike things, but users should
be able to load these tables without issue.

I also added a mapping for the `name` type, which shows up in some of
the default tables in `postgis`.