docs: update sankey diagram (#9077)
## Description of changes

this started as trying to get dynamic light/dark mode for plotting (with
the sankey diagram, but any) working with Quarto. I still couldn't
figure it out. there's definitely a path that's like save both versions
of the image and use JavaScript to dynamically load the right one.
another option, if we could get the HTML for the page in Python, would
be to use that

so anyway, this actually just updates the sankey diagram to put it in a
funciton to choose light/dark mode, minor code improvements, and
improving the colors

one suggestion from Jim was simplifying the cateogires and I agree, so
now it's just SQL/DataFrame

## Issues closed