docs(udfs): improve udf docstring examples (#9079)
UDF arguments need to be Ibis expressions. In the existing docstring
examples
for the UDF decorators, it can be unclear that the integers being passed
to
`add_one` are implicitly converted to Ibis literals.

To make it clearer, I've added memtables to all the examples, and used
pandas-
and pyarrow-specific functions to the examples.