refactor(udf): remove hierarchical usage of schema (#9078)
I did not deprecate `schema` as we have elsewhere, because I don't think we can determine the quoting behavior of the backend at UDF definition time, which makes accepting dotted string paths harder. 

BREAKING CHANGE: The `schema` parameter for UDF definition has been removed. A new `catalog` parameter has been added. Ibis uses the word database to refer to a collection of tables, and the word catalog to refer to a collection of databases. You can use a combination of `catalog` and `database` to specify a hierarchical location for the UDF.