fix(to_sql): use default backend for sql generation when set
We weren't passing `use_default=True` when looking for backends in
`to_sql` when no `dialect` was specified.  This worked for nearly all
cases, except when an expression was built from a `memtable`.
A `memtable` doesn't have a specific backend, but if a user has set a
default backend, we should assume that dialect for rendering
sql (instead of our default of DuckDB sql)