feat(exasol): allow passing in-memory data to create_table
I don't know that we can unregister `_clean_up_tmp_table` for specific
tables, so Exasol might throw some atexit errors (which are ignored) at
shutdown, because it's attempting to drop tables that have already been
dropped (also not sure why Exasol complains about this with
`force=True`).

Still, I think it's better to not pollute the table-space with copies of
memtables for every table we create.