fix: fix and improve shape inference in many ops
Usually when using eg duckdb's REGEXP_REPLACE() function, the pattern and replacement are always scalar
constants. But actually it can accept columnar for all three of it's arguments.

These sort of assumptions appeared to be all over the place in the code base. I found these fixs by grepping for "shape_like"
and manually looking at all the instances.

I didn't add any test cases, that felt like a monumental thing
to do I wasn't willing to put the time in for. But IDK,
I'm not sure I would even want to put in all the tests, it would be SO much boilerplate.

To deal with Nodes sometimes having optional args,
I modified shape_like() to be more flexible.
IDK, that was sort of a lazy approach, we could be
much more verbose and have each op do this filtering individually, but I didn't that was worth it.