feat(sigcheck): first cut of signature checking utilities
This is a fork of some of the utilities in Scott Sanderson's
`python-interface`
project (https://github.com/ssanderson/python-interface).

While the upstream is no longer maintained, the goal of that project
aligns quite well with some of the issues we face with maintaining
consistent interfaces across backends.

I'm grabbing a few of these utilities and ripping out Python 2 support,
then I'm going to work on setting up tests for making sure that backend
methods match the signatures in the parent `BaseBackend` class.