test: check that empty structs are disallowed
See https://github.com/ibis-project/ibis/pull/9310 for a discussion of whether
empty structs should be allowed or disallowed.
We settled on disallowing them for now,
since pyarrow is like the only backend to support them,
and it is definitely easier to change our mind and allow them later rather than
disallow them later.

Perhaps if you stay in ibis-land, emptry structs are useful ie for doing type manipulations, or maybe you
only use them for intermediate calculations?
But until we see this concrete use case, don't worry about it.

I had to adjust the dask testcase generation because this change revealed a bug in our existing dask tests. `ibis.backends.dask.Backend.get_schema()` uses a method of `PandasData.infer_table(df.head(1))`, so we were infering the type of the `map_of_*` testing column as
`struct<>` since the first element in that test array of values was `{}`.
So the simple solution was to just put the "representative" value first. This doesn't really
solve the root problem in how dask does type inference, we probably want to improve this.
I'd guess we should infer the column type from a larger sample of values. 1 seems really absurdly small. I tried doing this but it was trickier than I would have thought,
with a head(10) I was only getting two values back, even thought there should be 3??
IDK that should get solved in a different PR.