docs(algolia): add methods from backend pages to algolia index
The backend-specific methods are added to the backend pages
programmatically using `quarto`.  That seems to lead to them not showing
up in our Algolia search, because they aren't being scraped.

It's a little gross, but I've added in a section to our api docs
rendering scripts to also dump the corresponding records for each method
into a backend-specific json file during `quarto` rendering.

Then those backend-specific records files are loaded and added to the
Algolia index in `upload-algolia-api.py`