fix(case): fix dshape, error on noncomparable and empty cases
This is pinning down the expected behavior for cases before tackling
the case() to cases() switch in
https://github.com/ibis-project/ibis/pull/9096
so that PR can be simpler

I move the validation for comparable-ness down into the operation so that
the logic is consolidated to one place.
in #9096 there might be multiple places that construct an ops.SimpleCase, and we don't want
to have to implement the validation in all
calling locations.

We could consider relaxing the limitation for non-empty cases later, but for now lets be strict.

I already fixed the shape of ops.SearchedCase in https://github.com/ibis-project/ibis/pull/9334,
but it looks like in that PR I forgot to also fix ops.SimpleCase, so I do that fix here.