fix(pyspark): try `USE [CATALOG|DATABASE]` before pyspark api
The unity catalog on Databricks has weird permissioning where it doesn't
allow (at least one) user to run `setCurrentCatalog` or `setCurrentDatabase`.

It _does_ allow setting those via `USE CATALOG mycat;` and `USE DATABASE
mydb;`.

This, however, is not part of standard Spark SQL and is special
Databricks Spark SQL.

SO, we try the special databricks SQL, if that throws a parser error, we
fall back to using the pyspark API methods.