test: remove `broken` marker, better define `notimpl` and `notyet` (#9688)
inspired by https://github.com/ibis-project/ibis/issues/9535

The broken mark had the meaning of "this broke for some reason, but it
wasn't the fault of this PR, so lets ignore it for now". I find this
only meaningful in the context of that individual PR though: as soon as
the PR is merged, the usefulness drops to near 0. I come across this
mark as I'm reading code and I have no idea what action I should take.
notimpl, notyet, and never are much more useful, they signal where the
fault lies and what to do about it.

I went through and converted all the instances of broken to one of these
other 3 marks. It was fairly fast and definitely error prone. I could
see the argument for converting these all to plain xfails when we aren't
sure. I'm not sure how bad it is for these marks to be incorrect.

I tweaked the prose that describes these 3 marks. I would love your
editing of them.